Added GridStructure and Column Parser
GridStructure provides a neat and easy way for the developer to add
properties for each column. The class discovers the columns in the
LinqToSql class provided to it. In the future I plan to introduce a
interface for providing data as well as using the present functionality
of using the LinqToSql class.

ColumnParser as the name suggests parses the names of the column and
applies rules. These rules basically identify the Id columns and breaks
up the column names based on capitals.